entity framework introduction

+++ Service/IPersonService.cs
@@ -16,12 +16,12 @@ namespace LexiconMvc.Service
 
         void DeleteByPhoneNumber(string phoneNumber);
 
-        void DeleteById(long id);
+        void DeleteById(int id);
 
         Person GetByPhoneNumber(String phoneNmber);
 
         PersonViewModel CreatePersonViewModel(Person person);
 
-        Person GetById(long id);
+        Person GetById(int id);
     }
 }
